refactor: separated class logic to separate files
All classes now occupy their own files and are linked to their dependencies in their files rather than in one central place.
